Preserve field ordering by adding placeholder values for deferred fields
During execution if a field returns a future the value doesn't get
assigned to the results dictionary until after that future is resolved.
So if there is a mix of deferred and non-deferred fields in a query the
non-deferred fields will always get returned first regardless of the
request ordering.

This PR fixes this issue by adding a placeholder to the result object
that gets overridden when the future gets resolved. That way field
ordering is preserved.

Fixes #1

Diff-Id: 94549