Allow chaining together multiple dataloader calls
This PR changes how the dataloader batching functionality works. Rather
than trying to "detect" if a future has a "deferred_callback" in the
execution context, we create a singleton `DataLoaderBatchCallbacks`
which all dataloaders add their `dispatch_queue` functions to when
needed. We can then run all the callbacks in the execution context to
complete the SyncFuture's. This allows us to chain dataloaders together.

Fixes #6

Diff-Id: daffd